App is now able to get weather from API and display correct information onto screen

+++ app/src/main/java/com/marty/dang/polarpointsweatherapp/data/model/CurrentWeatherModel.kt
@@ -8,12 +8,9 @@ import com.squareup.moshi.Json
  *   Copyright @ 2019 Dang, Marty. All rights reserved.
  */
 
-// @json in Moshi = @SerializedName in Gson
-
-
 data class CurrentWeatherModel(
-    @field:Json(name = "lat") val lat: Int? = null,
-    @field:Json(name = "lon") val lon: Int? = null,
+    @field:Json(name = "lat") val lat: Double? = null,
+    @field:Json(name = "lon") val lon: Double? = null,
     @field:Json(name = "timezone") val timezone: String? = null,
     @field:Json(name = "timezone_offset") val timezoneOffset: Int? = null,
     @field:Json(name = "current") val current: CurrentModel? = null
